myapp/ml-service/requirements.txt
# Bu dosya Flask tabanli ML servisinin Python bagimliliklarini tanimlar.
Flask==3.1.0
mysql-connector-python==9.2.0
numpy==2.0.2
scikit-learn==1.6.1
tensorflow==2.18.0
python-dotenv==1.0.1
